audio_transcriber/requirements.txt
openai>=1.30.0
openai-whisper>=20231117
sounddevice>=0.4.6
soundfile>=0.12.1
numpy>=1.24.0
typer[all]>=0.9.0
python-dotenv>=1.0.1
rumps>=0.4.0
